tested storage

+++ moneycop/storage.py
@@ -2,5 +2,5 @@ import datetime
 from typing import List
 
 class Storage():
-    def save(self, amount: int, location: str, date: datetime) -> List:
+    def save(self, amount: float, location: str, date: datetime) -> List:
         raise NotImplementedError
